Update actions/checkout to v3

+++ src/BinaryTree.cpp
@@ -13,10 +13,10 @@ namespace CppSampleCode
   bool isBalance(const BinaryTree_p &tree, const Node_p &node)
   {
 
-    int32_t leftHeight = tree->getHeight(node->left);
-    int32_t rightHeight = tree->getHeight(node->right);
+    size_t leftHeight = tree->getHeight(node->left);
+    size_t rightHeight = tree->getHeight(node->right);
 
-    if (std::abs(leftHeight - rightHeight) > 1)
+    if (std::abs(static_cast<int64_t>(leftHeight - rightHeight)) > 1)
     {
       return false;
     }
